+++ src/Models/Round.php
@@ -19,6 +19,7 @@ use Illuminate\Database\Eloquent\Builder;
  * @property bool $mixed_alignment
  * @property \Illuminate\Support\Carbon $start_date
  * @property \Illuminate\Support\Carbon $end_date
+ * @property \Illuminate\Support\Carbon $offensive_actions_prohibited_at
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
  * @property-read \Illuminate\Database\Eloquent\Collection|\OpenDominion\Models\Dominion[] $dominions
@@ -34,7 +35,13 @@ use Illuminate\Database\Eloquent\Builder;
  */
 class Round extends AbstractModel
 {
-    protected $dates = ['start_date', 'end_date', 'created_at', 'updated_at'];
+    protected $dates = [
+        'start_date',
+        'end_date',
+        'offensive_actions_prohibited_at',
+        'created_at',
+        'updated_at'
+    ];
 
     // Eloquent Relations
 
@@ -128,6 +135,20 @@ class Round extends AbstractModel
     public function hasEnded()
     {
         return ($this->end_date <= now());
+    }
+
+    /**
+     * Returns whether offensive actions (and exploration) are disabled for the
+     * rest of the round.
+     *
+     * Actions like these are disabled near the end of the round to prevent
+     * suicides and whatnot.
+     *
+     * @return bool
+     */
+    public function hasOffensiveActionsDisabled(): bool
+    {
+        return ($this->offensive_actions_prohibited_at <= now());
     }
 
     /**