+++ src/Services/Dominion/Actions/SpellActionService.php
@@ -152,8 +152,16 @@ class SpellActionService
                 $result = $this->castInfoOpSpell($dominion, $spellKey, $target);
 
             } elseif ($this->spellHelper->isBlackOpSpell($spellKey)) {
+                if($dominion->round->hasOffensiveActionsDisabled())
+                {
+                    throw new GameException('Black ops have been disabled for the remainder of the round.');
+                }
                 throw new LogicException('Not yet implemented');
             } elseif ($this->spellHelper->isWarSpell($spellKey)) {
+                if($dominion->round->hasOffensiveActionsDisabled())
+                {
+                    throw new GameException('Black ops have been disabled for the remainder of the round.');
+                }
                 throw new LogicException('Not yet implemented');
             } else {
                 throw new LogicException("Unknown type for spell {$spellKey}");